requirements.txt
bokeh==2.0.1
bs4==0.0.1
dateparser==0.7.4
elasticsearch==7.5.1
Flask==1.1.1
https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-2.2.5/fr_core_news_md-2.2.5.tar.gz
lxml==4.5.0
pandas==1.0.1
python-dotenv==0.11.0
scrapinghub==2.3.1
Scrapy==2.0.1
shub==2.10.0
spacy==2.2.3
Unidecode==1.1.1
